+++ core/lib/src/constants.dart
