+++ core/lib/src/model/account_meta.dart
