+++ core/lib/src/model/app_status.dart
