+++ core/lib/src/model/geo_location.dart
