+++ core/lib/src/model/inapp/action.dart
@@ -1,4 +1,4 @@
-import 'package:moengage_flutter/model/inapp/inapp_action_type.dart';
+import 'inapp_action_type.dart';
 
 abstract class Action {
   ActionType actionType;
