+++ core/lib/src/model/inapp/campaign_context.dart
