+++ core/lib/src/model/inapp/click_data.dart
@@ -1,7 +1,7 @@
-import 'package:moengage_flutter/model/account_meta.dart';
-import 'package:moengage_flutter/model/inapp/action.dart';
-import 'package:moengage_flutter/model/inapp/campaign_data.dart';
-import 'package:moengage_flutter/model/platforms.dart';
+import '../account_meta.dart';
+import '../platforms.dart';
+import 'action.dart';
+import 'campaign_data.dart';
 
 class ClickData {
   Platforms platform;
