+++ core/lib/src/model/inapp/inapp_action_type.dart
