+++ core/lib/src/model/inapp/inapp_data.dart
@@ -1,6 +1,6 @@
-import 'package:moengage_flutter/model/account_meta.dart';
-import 'package:moengage_flutter/model/inapp/campaign_data.dart';
-import 'package:moengage_flutter/model/platforms.dart';
+import '../account_meta.dart';
+import '../platforms.dart';
+import 'campaign_data.dart';
 
 class InAppData {
   Platforms platform;
