+++ core/lib/src/model/inapp/navigation_type.dart
