+++ core/lib/src/model/inapp/self_handled_campaign.dart
@@ -1,5 +1,3 @@
-import 'package:moengage_flutter/constants.dart';
-
 class SelfHandledCampaign {
   /// Self handled campaign payload.
   String payload;
