+++ core/lib/src/model/inapp/self_handled_data.dart
@@ -1,7 +1,7 @@
-import 'package:moengage_flutter/model/account_meta.dart';
-import 'package:moengage_flutter/model/inapp/campaign_data.dart';
-import 'package:moengage_flutter/model/inapp/self_handled_campaign.dart';
-import 'package:moengage_flutter/model/platforms.dart';
+import '../account_meta.dart';
+import '../platforms.dart';
+import 'campaign_data.dart';
+import 'self_handled_campaign.dart';
 
 class SelfHandledCampaignData {
   CampaignData campaignData;
