+++ core/lib/src/model/permission_type.dart
