+++ core/lib/src/model/platforms.dart
