+++ core/lib/src/moe_ios_core.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/services.dart';
-import 'package:moengage_flutter/utils.dart';
-import 'package:moengage_flutter/model/gender.dart';
-import 'package:moengage_flutter/properties.dart';
-import 'package:moengage_flutter/model/app_status.dart';
-import 'package:moengage_flutter/constants.dart';
-import 'package:moengage_flutter/model/geo_location.dart';
 
+import 'constants.dart';
 import 'data_payload_mapper.dart';
+import 'model/app_status.dart';
+import 'model/gender.dart';
+import 'model/geo_location.dart';
+import 'properties.dart';
+import 'utils.dart';
 
 class MoEiOSCore {
   MethodChannel _channel;
